ADD ability to specify policy config yaml as URL
This commit adds the ability to specify an https URL for a policy config
YAML file.

Resolves: EC-40
