+++ internal/validate/helpers.go
@@ -29,10 +29,11 @@ import (
 
 // Determine policyConfig
 func GetPolicyConfig(ctx context.Context, policyConfiguration string) (string, error) {
-	// Check if policyConfiguration is a git url. If so, try to download a config file from git.
-	// If successful we write that into the data.policyConfiguration var.
-	if source.SourceIsGit(policyConfiguration) {
-		log.Debugf("Fetching policy config from git url %s", policyConfiguration)
+	// If policyConfiguration is not detected as a file and is detected as a git URL,
+	// or if policyConfiguration is an https URL try to download a config file from
+	// the provided source. If successful we read its contents and return it.
+	if source.SourceIsGit(policyConfiguration) && !source.SourceIsFile(policyConfiguration) || source.SourceIsHttp(policyConfiguration) {
+		log.Debugf("Fetching policy config from url: %s", policyConfiguration)
 
 		// Create a temporary dir to download the config. This is separate from the workDir usd
 		// later for downloading policy sources, but it doesn't matter because this dir is not
@@ -45,32 +46,38 @@ func GetPolicyConfig(ctx context.Context, policyConfiguration string) (string, e
 		defer utils.CleanupWorkDir(fs, tmpDir)
 
 		// Git download and find a suitable config file
-		configFile, err := source.GitConfigDownload(ctx, tmpDir, policyConfiguration)
+		configFile, err := source.GoGetterDownload(ctx, tmpDir, policyConfiguration)
 		if err != nil {
 			return "", err
 		}
-
-		// Changing data.policyConfiguration to the name of the newly downloaded file means we can
-		// use the code below to load the config
-		policyConfiguration = configFile
+		return readPolicyConfigurationFile(ctx, configFile)
+	} else if source.SourceIsFile(policyConfiguration) && utils.HasJsonOrYamlExt(policyConfiguration) {
+		// If policyConfiguration is detected as a file and it has a json or yaml extension,
+		// we read its contents and return it.
+		return readPolicyConfigurationFile(ctx, policyConfiguration)
 	}
 
-	// Check if policyConfig is a file path. If so, try to read it. If successful we write
-	// that into the data.policyConfiguration var.
-	if utils.HasJsonOrYamlExt(policyConfiguration) {
-		fs := utils.FS(ctx)
-		policyBytes, err := afero.ReadFile(fs, policyConfiguration)
-		if err != nil {
-			return "", err
-		}
-		// Check for empty file as that would cause a false "success"
-		if len(policyBytes) == 0 {
-			err := fmt.Errorf("file %s is empty", policyConfiguration)
-			return "", err
-		}
-		log.Debugf("Loaded %s as policyConfiguration", policyConfiguration)
-		policyConfiguration = string(policyBytes)
-	}
+	// If policyConfiguration is not a file path, git url, or https url,
+	// we assume it's a string and return it as is.
 	return policyConfiguration, nil
+
+}
+
+// Read policyConfiguration file and return its contents.
+func readPolicyConfigurationFile(ctx context.Context, policyConfiguration string) (string, error) {
+	// Check if policyConfig is a file path. If so, try to read it.
+	// If successful we write that into the data.policyConfiguration var.
+	fs := utils.FS(ctx)
+	policyBytes, err := afero.ReadFile(fs, policyConfiguration)
+	if err != nil {
+		return "", err
+	}
+	// Check for empty file as that would cause a false "success"
+	if len(policyBytes) == 0 {
+		err := fmt.Errorf("file %s is empty", policyConfiguration)
+		return "", err
+	}
+	log.Debugf("Loaded %s as policyConfiguration", policyConfiguration)
+	return string(policyBytes), nil
 
 }
